requirements.txt
aiogram>=3.15.0
anthropic>=0.40.0
python-dotenv>=1.0.0
aiosqlite>=0.20.0
tavily-python>=0.3.0
openai>=1.0.0
